Create service for booking price calculation
